:tractor: Reorganize a few functions around cobra bindings for clarity
